Exercise P2T3_area by figures geometrics
application that calculates the area of a circle, square or triangle. The user will be asked to enter the figure to which they want to calculate their area. As entered you will ask for the necessary values to calculate the area. Create a method for each figure to calculate each area, it will return a real number. Display the result on screen. Here we show you what each figure needs:
 Circle: (radius ^ 2) * PI
Triangle: (base * height) / 2
Square: side * side